Fix an intermittent send error unique to macOS
We are seeing EPROTOTYPE returned from send in our unit tests. According to http://erickt.github.io/blog/2014/11/19/adventures-in-debugging-a-potential-osx-kernel-bug this is a possible error in macOS when it probably should be returning an EPIPE.

Since we’re tearing the transport down in the EPIPE case, we just treat EPROTOTYPE as EPIPE in thise case.